Shut down running servers on plugin reload
This catches the plugin being unloaded and stops any running servers
that might exist in any current windows. Plugins reload when the
package they're in is disabled, and also when the plugin file itself is
modified while Sublime is running.

This change makes it easier to develop the plugin because you don't need
to restart Sublime every time you make a change, which would otherwise
leave servers running in the background with no way to stop them
because the connection between the command and the thread is severed.

This also sets up for being a package, in which case it's important that
disabling the package stops it from running.