Fix errors related to view activation
In the Sublime plugin ecosystem, a `view` represents an editable area of
text, which encompasses not only files that you have open but also the
edit components in things like the console or the project quick switch
panel.

In such cases, the `view` is not necessarily attached to a window, so
checks need to be done to make sure that the view has a window before
proceeding, or exceptions are thrown in the console.