Allow serving from windows with just open folders
In the original implementation, the location of the `sublime-project`
file loaded into the window was used as the root folder for the server.

This changes that so that in the absence of a project file, having at
least one folder open in the window is also an acceptable way to
configure the path. When there's more than one folder open, the first
one(at the top of the side bar) is the one used.

As an addition, the command now also disables itself if the config file
is not present in the root, whereas previously it would throw an
exception in that case.